moved FindVolumeInBundle
